// types/layout.d.ts
export interface MenuModal {
    label?: string;
    icon?: string;
    items?: MenuModal[];
    to?: string;
    url?: string;
    target?: HTMLAttributeAnchorTarget;
    separator?: boolean;
}